Update standard.md
a) Can the standards/best practices requirement be fleshed out more? I know it is supposed to be vague, but golly, anybody can check that box.
b) I split up Item 9, and put the privacy part into Item 7.
c) I'm personally not keen on the freedom of expression and human rights requirements -- I think generally they are very vague. That said, this is a group process, so I left that in.
d). Do no harm is now separate as Item 12 and is much less onerous. "Do no harm" is a very idealistic goal, but incredibly challenging to prove legally. This is not something that needs to be tied to years cases being heard at The Hague, Geneva, and the like.